water-meter/requirement.txt
automationhat==0.1.0
backports.functools-lru-cache==1.5
beautifulsoup4==4.7.1
blinker==1.3
blinkt==0.1.2
buttonshim==0.0.2
cachetools==3.1.1
Cap1xxx==0.1.3
certifi==2019.3.9
chardet==3.0.4
click==6.6
colorama==0.3.7
colorzero==1.1
cryptography==1.7.1
dill==0.2.9
drumhat==0.1.0
enum34==1.1.6
envirophat==1.0.0
ExplorerHAT==0.4.2
Flask==0.12.1
fourletterphat==0.1.0
funcsigs==1.0.2
future==0.16.0
futures==3.2.0
gax-google-logging-v2==0.8.3
gax-google-pubsub-v1==0.8.3
gcloud==0.17.0
google-api-core==1.11.0
google-auth==1.6.3
google-cloud-core==1.0.0
google-cloud-storage==1.16.0
google-gax==0.16.0
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
gpiozero==1.5.0
grpc-google-logging-v2==0.8.1
grpc-google-pubsub-v1==0.8.1
grpcio==1.21.1
httplib2==0.12.3
idna==2.8
ipaddress==1.0.17
itsdangerous==0.24
Jinja2==2.8
jws==0.1.3
keyring==10.1
keyrings.alt==1.3
MarkupSafe==0.23
mcpi==0.1.1
microdotphat==0.2.1
mock==2.0.0
mote==0.0.4
motephat==0.0.2
numpy==1.12.1
oauth2client==3.0.0
oauthlib==2.0.1
pantilthat==0.0.7
pbr==5.2.0
phatbeat==0.1.1
pianohat==0.1.0
picamera==1.13
picraft==1.0
piglow==1.2.4
pigpio==1.38
Pillow==4.0.0
ply==3.8
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycrypto==2.6.1
pycryptodome==3.4.3
pygame==1.9.3
pygobject==3.22.0
pyinotify==0.9.6
PyJWT==1.4.2
pyOpenSSL==16.2.0
Pyrebase==3.0.27
pyrfidhid==1.0
pyserial==3.2.1
python-jwt==2.0.1
pytz==2019.1
pyusb==1.0.2
pyxdg==0.25
rainbowhat==0.1.0
requests==2.22.0
requests-oauthlib==0.7.0
requests-toolbelt==0.7.0
RPi.GPIO==0.6.5
rsa==4.0
RTIMULib==7.2.1
scrollphat==0.0.7
scrollphathd==1.2.1
SecretStorage==2.3.1
sense-emu==1.1
sense-hat==2.2.0
simplejson==3.10.0
six==1.12.0
skywriter==0.0.7
sn3218==1.2.7
soupsieve==1.9.1
spidev==3.3
touchphat==0.0.1
transitions==0.6.9
twython==3.4.0
unicornhathd==0.0.4
urllib3==1.25.3
Werkzeug==0.11.15
